+++ pallets/attestation/src/lib.rs
@@ -47,7 +47,7 @@ mod tests;
 
 #[frame_support::pallet]
 pub mod pallet {
-    use entropy_shared::{AttestationQueue, KeyProvider, QuoteInputData};
+    use entropy_shared::{AttestationHandler, QuoteInputData};
     use frame_support::pallet_prelude::*;
     use frame_support::traits::Randomness;
     use frame_system::pallet_prelude::*;
@@ -77,10 +77,6 @@ pub mod pallet {
         type WeightInfo: WeightInfo;
         /// Something that provides randomness in the runtime.
         type Randomness: Randomness<Self::Hash, BlockNumberFor<Self>>;
-        /// A type used to get different keys for a given account ID.
-        type KeyProvider: entropy_shared::KeyProvider<Self::AccountId>;
-        /// A type used to describe a queue of attestations.
-        type AttestationQueue: entropy_shared::AttestationQueue<Self::AccountId>;
     }
 
     #[pallet::genesis_config]
@@ -119,6 +115,7 @@ pub mod pallet {
     #[pallet::generate_deposit(pub(super) fn deposit_event)]
     pub enum Event<T: Config> {
         AttestationMade,
+        AttestationIssued(Vec<u8>, BlockNumberFor<T>),
     }
 
     /// Errors related to the attestation pallet
@@ -140,6 +137,8 @@ pub mod pallet {
         CannotDecodeVerifyingKey,
         /// Could not verify PCK signature
         PckVerification,
+        /// There's an existing attestation request for this account ID.
+        OutstandingAttestationRequest,
     }
 
     #[pallet::call]
@@ -151,20 +150,69 @@ pub mod pallet {
         #[pallet::weight({
             <T as Config>::WeightInfo::attest()
         })]
-        pub fn attest(origin: OriginFor<T>, quote: Vec<u8>) -> DispatchResult {
+        pub fn attest(origin: OriginFor<T>, _quote: Vec<u8>) -> DispatchResult {
+            let _who = ensure_signed(origin)?;
+
+            Self::deposit_event(Event::AttestationMade);
+
+            Ok(())
+        }
+
+        /// Indicates to the chain that the caller wants to make an attestation.
+        ///
+        /// Once the chain is aware of this request, other extrinsics will be able to determine
+        /// whether or not the caller has provided a valid attestation.
+        #[pallet::call_index(1)]
+        #[pallet::weight({
+            <T as Config>::WeightInfo::request_attestation()
+        })]
+        pub fn request_attestation(origin: OriginFor<T>) -> DispatchResult {
             let who = ensure_signed(origin)?;
 
+            // We only want one pending attestation request per account.
+            ensure!(
+                !PendingAttestations::<T>::contains_key(&who),
+                Error::<T>::OutstandingAttestationRequest
+            );
+
+            let mut nonce = [0; 32];
+            Self::get_randomness().fill_bytes(&mut nonce[..]);
+            Self::request_quote(&who, nonce);
+
+            let block_number = <frame_system::Pallet<T>>::block_number();
+            Self::deposit_event(Event::AttestationIssued(nonce.to_vec(), block_number));
+
+            Ok(())
+        }
+    }
+
+    impl<T: Config> Pallet<T> {
+        fn get_randomness() -> ChaCha20Rng {
+            let phrase = b"quote_creation";
+            // TODO: Is randomness freshness an issue here
+            // https://github.com/paritytech/substrate/issues/8312
+            let (seed, _) = T::Randomness::random(phrase);
+            // seed needs to be guaranteed to be 32 bytes.
+            let seed = <[u8; 32]>::decode(&mut TrailingZeroInput::new(seed.as_ref()))
+                .expect("input is padded with zeroes; qed");
+            ChaChaRng::from_seed(seed)
+        }
+    }
+
+    impl<T: Config> entropy_shared::AttestationHandler<T::AccountId> for Pallet<T> {
+        fn verify_quote(
+            attestee: &T::AccountId,
+            x25519_public_key: entropy_shared::X25519PublicKey,
+            provisioning_certification_key: entropy_shared::BoundedVecEncodedVerifyingKey,
+            quote: Vec<u8>,
+        ) -> Result<(), DispatchError> {
             // Check that we were expecting a quote from this validator by getting the associated
             // nonce from PendingAttestations.
             let nonce =
-                PendingAttestations::<T>::get(&who).ok_or(Error::<T>::UnexpectedAttestation)?;
+                PendingAttestations::<T>::get(attestee).ok_or(Error::<T>::UnexpectedAttestation)?;
 
             // Parse the quote (which internally verifies the attestation key signature)
             let quote = Quote::from_bytes(&quote).map_err(|_| Error::<T>::BadQuote)?;
-
-            // Get associated x25519 public key from staking pallet
-            let x25519_public_key =
-                T::KeyProvider::x25519_public_key(&who).ok_or(Error::<T>::NoX25519KeyForAccount)?;
 
             // Get current block number
             let block_number: u32 = {
@@ -174,7 +222,7 @@ pub mod pallet {
 
             // Check report input data matches the nonce, TSS details and block number
             let expected_input_data =
-                QuoteInputData::new(&who, x25519_public_key, nonce, block_number);
+                QuoteInputData::new(attestee, x25519_public_key, nonce, block_number);
             ensure!(
                 quote.report_input_data() == expected_input_data.0,
                 Error::<T>::IncorrectInputData
@@ -185,9 +233,6 @@ pub mod pallet {
                 .map_err(|_| Error::<T>::BadMrtdValue)?;
             let accepted_mrtd_values = pallet_parameters::Pallet::<T>::accepted_mrtd_values();
             ensure!(accepted_mrtd_values.contains(&mrtd_value), Error::<T>::BadMrtdValue);
-
-            let provisioning_certification_key =
-                T::KeyProvider::provisioning_key(&who).ok_or(Error::<T>::NoPCKForAccount)?;
 
             // Check that the attestation public key is signed with the PCK
             let provisioning_certification_key = decode_verifying_key(
@@ -202,47 +247,15 @@ pub mod pallet {
                 .verify_with_pck(provisioning_certification_key)
                 .map_err(|_| Error::<T>::PckVerification)?;
 
-            PendingAttestations::<T>::remove(&who);
-            T::AttestationQueue::confirm_attestation(&who);
+            PendingAttestations::<T>::remove(attestee);
 
             // TODO #982 If anything fails, don't just return an error - do something mean
 
-            Self::deposit_event(Event::AttestationMade);
-
             Ok(())
         }
-    }
 
-    #[pallet::hooks]
-    impl<T: Config> Hooks<BlockNumberFor<T>> for Pallet<T> {
-        fn on_initialize(now: BlockNumberFor<T>) -> Weight {
-            let pending_validators = T::AttestationQueue::pending_attestations();
-            let num_pending_attestations = pending_validators.len() as u32;
-            let mut requests = AttestationRequests::<T>::get(now).unwrap_or_default();
-
-            for account_id in pending_validators {
-                let mut nonce = [0; 32];
-                Self::get_randomness().fill_bytes(&mut nonce[..]);
-                PendingAttestations::<T>::insert(&account_id, nonce);
-                requests.push(account_id.encode());
-            }
-
-            AttestationRequests::<T>::insert(now, requests);
-
-            <T as Config>::WeightInfo::on_initialize(num_pending_attestations)
-        }
-    }
-
-    impl<T: Config> Pallet<T> {
-        pub fn get_randomness() -> ChaCha20Rng {
-            let phrase = b"quote_creation";
-            // TODO: Is randomness freshness an issue here
-            // https://github.com/paritytech/substrate/issues/8312
-            let (seed, _) = T::Randomness::random(phrase);
-            // seed needs to be guaranteed to be 32 bytes.
-            let seed = <[u8; 32]>::decode(&mut TrailingZeroInput::new(seed.as_ref()))
-                .expect("input is padded with zeroes; qed");
-            ChaChaRng::from_seed(seed)
+        fn request_quote(who: &T::AccountId, nonce: [u8; 32]) {
+            PendingAttestations::<T>::insert(who, nonce)
         }
     }
 }