+++ CHANGELOG.md
@@ -11,9 +11,14 @@ At the moment this project **does not** adhere to
 
 ### Breaking Changes
 - In [#1104](https://github.com/entropyxyz/entropy-core/pull/1104) the `/validator/rotate_network_key` endpoint was renamed to `rotate_network_key`
+- In [#1109](https://github.com/entropyxyz/entropy-core/pull/1109/), the `MaxPendingAttestations` config type, the `ValidationQueue` storage
+  structure, and the `NodeInfoChanged` event were removed from the Staking Extension pallet. The
+  `AttestationHandler` config type was added to the Staking Extension pallet. The `KeyProvider` and
+  `AttestationQueue` config types were removed from the Attestation pallet.
 
 ### Changed
 - Use correct key rotation endpoint in OCW ([#1104](https://github.com/entropyxyz/entropy-core/pull/1104))
+- Change attestation flow to be pull based ([#1109](https://github.com/entropyxyz/entropy-core/pull/1109/))
 
 ## [0.3.0-rc.1](https://github.com/entropyxyz/entropy-core/compare/release/v0.2.0...release/v0.3.0-rc.1) - 2024-10-04